# openapi/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Gestionale Negozio API
  description: API del gestionale negozio
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/customers/{id}:
    get:
      tags:
      - customer-controller
      operationId: getCustomerById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      tags:
      - customer-controller
      operationId: updateCustomer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      tags:
      - customer-controller
      operationId: deleteCustomer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/colors/{id}:
    get:
      tags:
      - color-controller
      operationId: getColorById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColorResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      tags:
      - color-controller
      operationId: updateColor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateColorRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColorResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      tags:
      - color-controller
      operationId: deleteColor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/customers:
    get:
      tags:
      - customer-controller
      operationId: getAllCustomers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      tags:
      - customer-controller
      operationId: createCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/colors:
    get:
      tags:
      - color-controller
      operationId: getAllColors
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ColorResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      tags:
      - color-controller
      operationId: createColor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateColorRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColorResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/colors/search:
    get:
      tags:
      - color-controller
      operationId: searchColor
      parameters:
      - name: nome
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ColorResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/colors/by-name/{nome}:
    get:
      tags:
      - color-controller
      operationId: getColorByName
      parameters:
      - name: nome
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColorResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
    ApiError:
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ColorResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
          minLength: 1
      required:
      - id
      - nome
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
          minLength: 1
        cognome:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        indirizzo:
          type: string
        citta:
          type: string
        cap:
          type: string
        telefono:
          type: string
          minLength: 1
        telefonoSecondario:
          type: string
      required:
      - cognome
      - email
      - id
      - nome
      - telefono
...
    UpdateCustomerRequest:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        cognome:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        indirizzo:
          type: string
        citta:
          type: string
        cap:
          type: string
          pattern: "\\d{5}"
        telefono:
          type: string
          minLength: 1
        telefonoSecondario:
          type: string
      required:
      - cognome
      - email
      - nome
      - telefono
    UpdateColorRequest:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
      required:
      - nome
...
    CreateCustomerRequest:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        cognome:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        indirizzo:
          type: string
        citta:
          type: string
        cap:
          type: string
          pattern: "\\d{5}"
        telefono:
          type: string
          minLength: 1
        telefonoSecondario:
          type: string
      required:
      - cognome
      - email
      - nome
      - telefono
    CreateColorRequest:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
      required:
      - nome
...